modularize home screen
