+++ examples/amazonvega-demo/PurchaseTester/src/screens/ConfigureScreen.styles.ts
