+++ examples/amazonvega-demo/PurchaseTester/src/screens/HomeScreen.styles.ts
